Feature/Helpers Functions.
* Helper functions split into different modules.
