+++ connect/processors_toolkit/logger/__init__.py
@@ -6,4 +6,5 @@
 from .adapter import (  # noqa: F401
     bind_logger,
     ExtensionLoggerAdapter,
+    mask,
 )
