+++ connect/processors_toolkit/requests/__init__.py
@@ -8,9 +8,8 @@ from __future__ import annotations
 from copy import deepcopy
 from typing import Any, Dict, List, Optional, Union
 
-from connect.processors_toolkit import find_by_id, merge
 from connect.processors_toolkit.requests.assets import AssetBuilder
-from connect.processors_toolkit.requests.helpers import make_param, request_model
+from connect.processors_toolkit.requests.helpers import find_by_id, make_param, merge, request_model
 from connect.processors_toolkit.requests.tier_configurations import TierConfigurationBuilder
 from connect.processors_toolkit.requests.exceptions import MissingParameterError
