+++ connect/processors_toolkit/requests/helpers.py
@@ -3,37 +3,48 @@
 #
 # Copyright (c) 2022 Ingram Micro. All Rights Reserved.
 #
+from copy import deepcopy
 from typing import List, Optional, Union
 
 from faker import Faker
 
 
-def request_model(request: dict) -> str:
+def find_by_id(elements: List[dict], element_id: str, default: Optional[dict] = None) -> Optional[dict]:
     """
-    Returns the request model depending on the request type.
+    Searches for a parameter/item with the given ``id`` within the ``list``.
 
-    :param request: dict
-    :return: str
+    :param elements: The list of parameters/items to search.
+    :param element_id: The id of the parameter/item to find.
+    :param default: Default value to return if item is not found.
+    :return: The parameter/list, or ``default`` if it was not found.
     """
-
-    def match_request_type(model: dict) -> bool:
-        return model.get('object') in request or request.get('type') in model.get('types')
-
     try:
-        return next(filter(match_request_type, [
-            {
-                'request': 'asset',
-                'object': 'asset',
-                'types': ['adjustment', 'purchase', 'change', 'suspend', 'resume', 'cancel'],
-            },
-            {
-                'request': 'tier-config',
-                'object': 'configuration',
-                'types': ['setup'],
-            },
-        ])).get('request')
+        return next(filter(lambda element: element['id'] == element_id, elements))
     except StopIteration:
-        return 'undefined'
+        return default
+
+
+def merge(base: dict, override: dict) -> dict:
+    """
+    Merge two dictionaries (override into base) recursively.
+
+    :param base: The base dictionary.
+    :param override: Override dictionary to be merged into base.
+    :return dict: The new dictionary.
+    """
+    new_base = deepcopy(base)
+    for key, value in override.items():
+        if key in new_base:
+            if isinstance(new_base[key], dict) and isinstance(value, dict):
+                new_base[key] = merge(new_base[key], value)
+            elif isinstance(new_base[key], list) and isinstance(value, list):
+                new_base[key].extend(value)
+            else:
+                new_base[key] = value
+        else:
+            new_base[key] = value
+
+    return new_base
 
 
 def make_tier(tier_type: str = 'customer', locale: List[str] = None) -> dict:
@@ -85,3 +96,31 @@ def make_param(
         'scope': scope,
         'phase': phase,
     }
+
+
+def request_model(request: dict) -> str:
+    """
+    Returns the request model depending on the request type.
+
+    :param request: dict
+    :return: str
+    """
+
+    def match_request_type(model: dict) -> bool:
+        return model.get('object') in request or request.get('type') in model.get('types')
+
+    try:
+        return next(filter(match_request_type, [
+            {
+                'request': 'asset',
+                'object': 'asset',
+                'types': ['adjustment', 'purchase', 'change', 'suspend', 'resume', 'cancel'],
+            },
+            {
+                'request': 'tier-config',
+                'object': 'configuration',
+                'types': ['setup'],
+            },
+        ])).get('request')
+    except StopIteration:
+        return 'undefined'
